Fix NumPy array dtype to uint8 for PIL Image conversion

+++ hakim_ai/layer1_router/router_agent.py
@@ -55,7 +55,7 @@ class RouterAgent:
             # Simple feature-based classification on the thumbnail
             import numpy as np
             from PIL import Image
-            thumb_np = np.array(wsi_data.thumbnail)
+            thumb_np = np.array(wsi_data.thumbnail, dtype=np.uint8)
             patch_array = thumb_np[:224, :224] if thumb_np.shape[0] >= 224 and thumb_np.shape[1] >= 224 else thumb_np
             patch = Image.fromarray(patch_array)
